third_party/WebKit/LayoutTests/fast/dom/Window/Location/window-shadow-location-using-js-object-with-toString.html

<html>
<head>
    <script src="../../../../resources/js-test.js"></script>
    <script src="resources/location-tests-functions.js"></script>
</head>
<body>
<script>
    if (window.testRunner)
        testRunner.waitUntilDone();

    var result;
    window.onmessage = function()
    {
        result = normalizeURL(String(window.frames[0].location));
        shouldBe("result", "'pass.html'");
        if (window.testRunner)
            testRunner.notifyDone();
    }
</script>
<iframe src="resources/window-shadow-location-using-js-object-with-toString-iframe.html"></iframe>
</body>
</html>
